requirements.txt
celery
Django
django-cors-headers
flower
logzero
pandas
PyMySQL
pyotp
python-dateutil
pytz
redis
requests
smartapi-python
websocket-client
whitenoise
mysqlclient